organizing repo structure
organized backend files into backend folder and frontend files into frontend folder
