--- HashGuard/HashGuard/EventListModel.qml
@@ -1,12 +1,0 @@
-import QtQuick
-
-ListModel {
-    id: eventListModel
-
-    ListElement {
-        eventId: "enterPressed"
-        eventDescription: "Emitted when pressing the enter button"
-        shortcut: "Return"
-        parameters: "Enter"
-    }
-}
