--- HashGuard/HashGuard/designer/plugin.metainfo
@@ -1,13 +1,0 @@
-MetaInfo {
-    Type {
-        name: "HashGuard.EventListSimulator"
-        icon: ":/qtquickplugin/images/item-icon16.png"
-
-        Hints {
-            visibleInNavigator: true
-            canBeDroppedInNavigator: true
-            canBeDroppedInFormEditor: false
-            canBeDroppedInView3D: false
-        }
-    }
-}
