--- HashGuard/HashGuardContent/App.qml
@@ -1,16 +1,0 @@
-import QtQuick
-import HashGuard
-
-Window {
-    width: mainScreen.width
-    height: mainScreen.height
-
-    visible: true
-    title: "HashGuard"
-
-    Screen01 {
-        id: mainScreen
-    }
-
-}
-
