--- HashGuard/HashGuardContent/fonts/fonts.txt
@@ -1,1 +1,0 @@
-Fonts in this folder are loaded automatically.
