--- HashGuard/HashGuardContent/images/images.txt
@@ -1,1 +1,0 @@
-Default folder for image assets.
