+++ backend/config.yaml
@@ -1,0 +1,27 @@
+watch_path: ./downloads
+quarantine_path: ./quarantine
+min_size_bytes: 1024
+max_size_bytes: 2147483648
+allowed_extensions:
+  - .exe
+  - .msi
+  - .zip
+  - .7z
+  - .pdf
+  - .docx
+  - .js
+blocked_paths:
+  - /tmp
+stability_seconds: 3
+debounce_ms: 500
+worker_count: 4
+cache_db: ./cache.sqlite
+ignore_name_regex:
+  - '^~\\$'
+  - '\\.crdownload$'
+priority_map:
+  '.exe': 10
+  '.msi': 9
+  '.zip': 7
+  '.7z': 7
+  '.pdf': 5
