+++ frontend/HashGuard.qrc
