+++ frontend/HashGuard/qmldir
@@ -1,0 +1,4 @@
+module HashGuard
+singleton Constants 1.0 Constants.qml
+EventListSimulator 1.0 EventListSimulator.qml
+EventListModel 1.0 EventListModel.qml
