+++ frontend/HashGuardContent/App.qml
@@ -1,0 +1,16 @@
+import QtQuick
+import HashGuard
+
+Window {
+    width: mainScreen.width
+    height: mainScreen.height
+
+    visible: true
+    title: "HashGuard"
+
+    Screen01 {
+        id: mainScreen
+    }
+
+}
+
