+++ frontend/HashGuardContent/fonts/fonts.txt
@@ -1,0 +1,1 @@
+Fonts in this folder are loaded automatically.
