+++ frontend/HashGuardContent/images/images.txt
@@ -1,0 +1,1 @@
+Default folder for image assets.
