+++ frontend/qtquickcontrols2.conf
@@ -1,0 +1,6 @@
+; This file can be edited to change the style of the application
+; Read "Qt Quick Controls 2 Configuration File" for details:
+; https://doc.qt.io/qt-6/qtquickcontrols-configuration.html
+
+[Controls]
+Style=Basic
